Add providers hoc
