+++ website/frontend/src/context/alert/index.ts
@@ -1,0 +1,3 @@
+import { AlertProvider } from './Provider';
+
+export { AlertProvider };
